# Speedup `csvstat` by using a sniff-limit and faster `get_freq` implementation