--- a/csvkit/utilities/csvjson.py
+++ b/csvkit/utilities/csvjson.py
@@ -55,8 +55,9 @@ class CSVJSON(CSVKitUtility):
             '--stream', dest='streamOutput', action='store_true',
             help='Output JSON as a stream of newline-separated objects, rather than an as an array.')
         self.argparser.add_argument(
-            '-y', '--snifflimit', dest='sniff_limit', type=int,
-            help='Limit CSV dialect sniffing to the specified number of bytes. Specify "0" to disable sniffing.')
+            '-y', '--snifflimit', dest='sniff_limit', type=int, default=1024,
+            help='Limit CSV dialect sniffing to the specified number of bytes. '
+                 'Specify "0" to disable sniffing entirely, or "-1" to sniff the entire file.')
         self.argparser.add_argument(
             '-I', '--no-inference', dest='no_inference', action='store_true',
             help='Disable type inference (and --locale, --date-format, --datetime-format) when parsing CSV input.')
@@ -141,10 +142,11 @@ class CSVJSON(CSVKitUtility):
             self.argparser.error('--key is only allowed with --stream when --lat and --lon are also specified.')
 
     def read_csv_to_table(self):
+        sniff_limit = self.args.sniff_limit if self.args.sniff_limit != -1 else None
         return agate.Table.from_csv(
             self.input_file,
             skip_lines=self.args.skip_lines,
-            sniff_limit=self.args.sniff_limit,
+            sniff_limit=sniff_limit,
             column_types=self.get_column_types(),
             **self.reader_kwargs
         )